Added .gitignore file and removed example_resume.pdf, README.md, and resume.tex
Added a .gitignore file to ignore unnecessary files such as example_resume.pdf, README.md, and resume.tex. These files are specific to individual users and should not be included in the repository. This commit ensures that only essential files for the LaTeX resume template are tracked and committed to the repository.